Feature/relationships (#1)
Adds relationship functions to enable Resources to declare relationships with other resources. 

The method builders in the `helium.relations` module will inject the right methods in the source and target classes based on the arguments provided. 

This includes writable support which allows, for example, labels to have sensors added and removed to them.