Change create/update signatures
Resource and metadata creation and updating has changed to no longer take keyword arguments as the "dictionary" for attributes. They now take a attributes keyword argument that is a dictionary of the attributes.

This allows for (1) full string attribute keys and (2) additional arguments to those methods for needed arguments like relationships. See `Label.create` for an example